comment in Telnet:
sorta kinda works.  The Consumer connects to the remote telnet client
correctly and prints the output.  Also, when the NoSuchElementException
log level is set to info or higher it will spam the console with output.
this indicates that the Consumer infinite while loop is excecuting.

The Producer works fine as an infinite loop, it correctly echoes back user
input to the console.

However, probably the program gets "stuck" in the Producer loop.

It's hard to debug because the options are either not enough info, or way
too much spamming from producer.